requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
python-jose
passlib[bcrypt]
